Treat pinned keys as bindings that will exist.
When we introduced "pins" to handle cases such as "bind(Foo.class);" that didn't
actually introduce a binding, I missed one case.  Specifically, if a child
module needs Foo before we've visited the parent and created the binding for the
Pin, the child should still inherit the binding from the parent, because the Pin
is a promise to create a binding.  If we don't respect that, we actually end up
in a state where we're "promising to create a double binding".  And then we do.

git-svn-id: http://google-gin.googlecode.com/svn/trunk@228 1a334de8-7455-0410-9b2e-eb0121b2e18b