Detect double bindings after reification.
When we yanked out the "Validate w/ Guice" logic, we did so because any double
bindings introduced during resolution were already detected by the resolution
algorithm.  This was correct... however, we also lost the ability to detect
double bindings from the user, and the double binding detection served as a nice
sanity check.  This CL walks the dependency tree (after reification) and
verifies that there are no double bindings.

git-svn-id: http://google-gin.googlecode.com/svn/trunk@229 1a334de8-7455-0410-9b2e-eb0121b2e18b