Don't treat implicit bindings as conflicting with themselves (by dburrows)
When Gin discovers unresolved dependencies, it checks that they don't exist in
child modules in order to avoid double bindings.  Unfortunately, ParentBindings
that pointed at pins were triggering this check, even though they just pointed
at the "slot" where a binding was about to be created!

This led to a situation where a binding was rejected with an error because it
conflicted with itself.  Added a test for this case and fixed the code.

git-svn-id: http://google-gin.googlecode.com/svn/trunk@232 1a334de8-7455-0410-9b2e-eb0121b2e18b